tests/test_human_in_the_loop.py
import unittest
from unittest import mock  # Import the mock library
from safety.human_in_the_loop import HumanInTheLoop

class TestHumanInTheLoop(unittest.TestCase):
    def setUp(self):
        self.hitl = HumanInTheLoop()
    
    def tearDown(self):
        # Your cleanup code here. For example:
        self.hitl = None

    def test_execute_function_with_hitl_whitelisted(self):
        with mock.patch('builtins.input', return_value='y'):  # Mocking input to return 'y'
            result = self.hitl.execute_function_with_hitl("WhitelistedFunc", "test_user")
        self.assertEqual(result, "Function Executed")  # Update this to the actual returned value

    def test_execute_function_with_hitl_user_permissions(self):
        with mock.patch('builtins.input', return_value='y'):  # Mocking input to return 'y'
            result = self.hitl.execute_function_with_hitl("UserSpecificFunc", "test_user")
        self.assertEqual(result, "Function Executed")  # Replace this with the expected result

    def test_execute_function_with_hitl_denied(self):
        with mock.patch('builtins.input', return_value='n'):  # Mocking input to return 'n'
            result = self.hitl.execute_function_with_hitl("DeniedFunc", "test_user")
        self.assertEqual(result, "Command not approved")

    def test_human_approval(self):
        # TODO: Write a test to ensure the human_approval mechanism works as expected.
        pass  # Placeholder, implement this test as needed

if __name__ == '__main__':
    unittest.main()
